Warnings removed. Fixed a nil datasocket value on cmd_store_tempfile

+++ lib/em-ftpd/files.rb
@@ -128,26 +128,29 @@ module EM::FTPD
       tmpfile.binmode
 
       wait_for_datasocket do |datasocket|
-        datasocket.on_stream { |chunk|
-          tmpfile.write chunk
-        }
-        send_response "150 Data transfer starting"
-        datasocket.callback {
-          puts "data transfer finished"
-          tmpfile.flush
-          @driver.put_file(target_path, tmpfile.path) do |bytes|
-            if bytes
-              send_response "226 OK, received #{bytes} bytes"
-            else
-              send_action_not_taken
+        # datasocket may be nil if connection get timeout'ed
+        if datasocket
+          datasocket.on_stream { |chunk|
+            tmpfile.write chunk
+          }
+          send_response "150 Data transfer starting"
+          datasocket.callback {
+            puts "data transfer finished"
+            tmpfile.flush
+            @driver.put_file(target_path, tmpfile.path) do |bytes|
+              if bytes
+                send_response "226 OK, received #{bytes} bytes"
+              else
+                send_action_not_taken
+              end
             end
-          end
-          tmpfile.unlink
-          close_datasocket # force closing the datasocket
-        }
-        datasocket.errback {
-          tmpfile.unlink
-        }
+            tmpfile.unlink
+            close_datasocket # force closing the datasocket
+          }
+          datasocket.errback {
+            tmpfile.unlink
+          }
+        end
       end
     end
